Have the mission to not kill pug defenders use `save` instead of `kill`
Right now the Wanderer reaction to killing the pug during
"Wanderers Ap'arak 3: Pug" only activates on landing on "Varu Tev'kei".
It also allows you to capture one of the Arfectas and continue the
Wanderer campaign. This makes it so that the mission is resolved upon
landing anywhere, and it fails the Wanderer campaign even if you
capture or destroy just one of them.