Embed test-savegame data in test-data-files
Embed the savegame data for tests in the files containing the other
test-data. This allows plugins to also supply savegames for usage during
testing.

Another benefit is that we are one step closer to usage of temporary
directories for running auto-tests (the directories don't need to be
filled up-front with savegame data, since the test itself writes out the
data before the game needs it).

Signed-off-by: Peter van der Meer <peter.vd.meer@gmail.com>