Add formation flying functionality
Added:
* Classes describing Formations
* Gamedata entries
* AI handling of formations
* Setting, loading and saving of formations on Ships
* Some example formations (data-file).

UI functionality (to set the formations through the UI) is not yet implemented.
Already sending this PR to allow people to experiment with formations (through datafile modifications or savegame editting).
Simplest way to experiment at this moment is to set a formation in data/ship.txt for ship types and then buy ships with the formation set.

Signed-off-by: Peter van der Meer <peter.vd.meer@gmail.com>