Guard positions
Add guard (position) functionality to allow escort ships to do station-
keeping on a relative offset to their lead ship.

The position is implemented as a child section of guard. This would be
overkill if we only want to set a position, but I envisioned that the
guard position functionality can be extended with an anchor that also
allows guard positions relative to:
* planet "<name>" - a specific planet - for example a ring of missile boats around a planet
* planet - any planet- for example for generic planetary defense fleets
* flagship - the players flagship - for example for NPCs following the player
* ship "<name>" - a specific ship
* local-system - for ships that should always take a fixed position in any system they travel to
* lead - The lead ship for escorts (current anchor, default anchor)

And maybe even for positions relative to:
* an incomming starlane - for example for pirates setting an ambush
* a wormhole - same example pirates, but now near a wormhole

Signed-off-by: Peter van der Meer <peter.vd.meer@gmail.com>