Randomly generate the offer dialouge for Courier 2
I'm not sure about the formatting, since I don't know of any examples of this in the current game.

This makes Courier 2 (the small package mission) have several different offer dialogues, which are generated in the same way as the hail dialogue. For an example of what this can generate: `A resident of Earth approaches you and says, "Sorry to interrupt, but I have a small package that needs to get to New Boston by April 32nd, 3012. I'm willing to pay you 2 credits for it."`. I also made it so that it tells you how much you'll get paid for it 3 out of 4 times. (The current offer never states payment.) The longest possible combination is 237 characters long, which is 43 characters shorter than Courier 3's offer.