Prepped the requiredOutfits loop to check specific locations
Did this based off of an optimized version of my pseudo code. What I have so far does compile and I expect it would run in game, but change the bevhaior for require to ALWAYS check EVERY ship in the player's fleet.
This is not functional, and requires some of the TODO comments to be finished before being functional.
Some of the TODO comments are questions about this implemenation.