Mark II Frigate given matching visuals.
Currently, the Mark II Frigate is the only Mark II ship in the game to still use ion thrusters. This causes mismatching visuals which make it look out of place compared to the other Mark II ships. This PR gives it a250 Atomic thrusters, matching with the rest of the ships and making it move around 65-70 speed faster. This makes it slightly more swift but still slower than a dreadnought. The Frigate is also given extra energy storage with the extra outfit space.

Suggested by @Arachi-Lover