Knocked the Mule down a bit as to be less of a master of all trades
Reduction in cargo space means the Mule now has less cargo space than every human heavy freighter. 200 is also a midpoint between the highest light freighter cargo and least heavy freighter cargo, so this now bridges a gap instead of being a no-brainer over some heavy freighters.

Increase in required crew may be overboard at this point, but better safe than sorry. 6 was far too low compared to all the other MWs.

Reduction in weapon space so that this is no longer the MW with the most weapon space. -20 is due to the fact that the Mule carries a figher that's basically worth +20 weapon space, and  an extra -10 was done for safe measure. The weapon space could possible be reduced further without harming the integrity of the ship, but further reductions would mean needing to change the stock loadout and the loadouts of more variants.

Reduction in outfit space by -10 to differentiate from the Osprey.

Updated the description to make it clear that this is only a jack of all trades, and dedicated ships would be better.