Slight tweaks to the Osprey to make it a more viable choice
Reduced the hull cost as to make it more appealing. The stock cost will still be the most among human MWs by quite a bit.

Reduced the drag slightly to make it a faster ship that's more competitive with the Corvette in terms of speed, but the Corvette is still able to be faster by quite a bit when using a reasonable engine loadout.