Improved some of the Bastion's stats to make it more of a tank
The Bastion is the tank among the MWs, but the HP didn't actually make up for how slow the ship is. Increased the shields and hull so that the Bastion more notably stands out among the MWs instead of only having slightly more HP than the other high HP MWs.

Increased the weapon capacity slightly to differentiate it from the Osprey and make it fit the tank role better.

Increased the cost slightly to compensate for the improved stats.