Tweaking Pirate Descriptions
Changed the description for Haven and Smuggler's Den to include a mention that customized ships are available to those with the right connections. Smugler's Den's description also has a comment added about it being a hub for fenced goods from beyond the area.

Also added "Lovelace Basics" outfitter to Smuggler's Den.