Creating the Cassette Class Engine
This engine is intended to be a smaller and more compact engine unit for the Tern (a Remnant drone), in order to get both the drone and its available space down.

This engine is a combination thruster-steering unit that takes up aproximately 75% of the space of a crucible-class thruster & steering unit combined (rounded up), and provides roughly 75% of the thrust & steering (rounded down). As such, the intention is that when compared to other thrusters or steering, it should be slightly less space effecient than other Remnant engines. (although only slightly)

Note: Due to the gascraft archon dodging mission specs that are intended to make using a cloaking device to bypass that mission impossible to beginner players, this PR also makes the following changes to the gascraft:
 - Replaces the Crucible-Class thruster & steering with a Cassette-Class engine
 - Adjusts the mass & drag of the gascraft to keep the Max Speed and acceleration as close to the original as possible. (<1% variation).
- Removes outfit space & engine space to reflect the smaller propulsion system.