More Varried Pirate Shipyards
This push splits the pirate shipyards into six distinct shipyards:
Southern, Northern, and Core sets, each with a normal and advanced version.  There is no overlap between the shipyards, so if a port is to have a full set of one of those, it needs to list both. Ex. "Southern Pirates" and "Advanced Southern Pirates".

No pirate shipyard has all the ships. Smugler's Den probably is the most advanced of them, having two of the "advanced" shipyards (Core and Southern)