Re-applying Amazinite's changes
Long story short, I have two void-sprite related mission branches going right now: One that is just the Face to Maw PR, and a second that builds on that for a larger mission chain. One way or another, I applied a large number of Amazinite's fixes to the second branch (that is not PR as it is nowhere close to ready yet) instead of to this one.

This push should get things back to where they should be.