Housekeeping
Don't sent a lone, undefended Auxiliary to take down some Alphas, even if the troop number is higher.

Fixed a reference to the Bactrian even though it's now an Auxiliary.

Lowered the drag on the Dropship and changed the descriptions of both new ships.

Removed Republic Logistics from the map for now.