Tweaked the Ibis to look more like a Remnant medium warship
Looking at the Albatross as an example, the in-built attributes of Remnant ships means that their outfit space is less than ships of comparable size form other factions. The Ibis' space was equal to or greater than similar ships though. These changes reign the Ibis in to better fit as a Remnant ship in the same way the Albatross does.