Expanding the tribute changes to the FWC side
Should be fairly similar to the Reconciliation route, just on the FWC side. When pug take over, the pug stop paying the tribute (one can assume that the measily 1 credit a day is either the Pug intentionally insulting the player by giving them a single credit in tribute payment; or some holdover in the banking system. Or maybe a desperate citizen keeping the payment alive to remind the player they are there.

Despite the lack of tribute payment, the pug will still actively respond to anyone challenging them, and send out fleets to destroy any who so dare.

After the liberation the defence fleets return to regular human ones - albeit slightly larger than before.