Adding more pirate names
A major goal was to add more female and non-western pirates, as well as make it easier to determine whether the pirate names were real or fictional. I realized a lot of the ones previously listed were ghost ships, so I moved them into their own category (perhaps if derelict ships are implemented, those names can be repurposed for that). I decided not to use any historical names newer than the last hundred years or so since piracy is still a modern problem.