Boosting the Strong Wind
The Strong Wind is described as a science vessel that has been retrofitted for combat. Given the Wanderer's focus on terraforming activities, it makes sense to me that their science vessel would have very powerful sensors related to that.

This PR adds the following to the Strong Wind:
- "asteroid scan power" 100
- "atmosphere scan" 100
- "tactical scan power" 20

This could result in the Strong Wind being an interesting candidate for a late-game mining ship for players so inclined. This change shouldn't have much impact on combat, but will help the Strong Wind feel a bit more unique than "just another warship."