Enhance the calculation process for crew salaries
Context:

Fleet management is currently very simple from an economic perspective.
Crew members cost a flat 100 credits per day, regardless of what they
actually do on their ship. Salary costs can be mostly ignored because
the ships' benefits to the player far outweigh them.

Parked ships are another oddity. The player can unpark them from any
planet and they will launch and jump to the player's system. This
implies that they actually have a crew ready and waiting for their fleet
captain's instructions. At the moment, the player doesn't have to pay
the crews of their parked ships at all. Somehow, the crew is magically
re-hired when the player presses the Unpark button.

Changes:

This commit makes crew salaries more complex by introducing two more
ranks for crew members: `commander` and `officer`.

- Every ship, regardless of size, requires one commander. We have to pay
each commander a salary of 1000 credits per day.

- For every 5 crew members that we have on a ship, we need an officer.
This rounds down; for example, a ship with 9 crew members needs only 1
officer but a ship with 10 crew must have 2. We must pay officers a
salary of 250 credits per day.

Considerations:

The goal of this change is to make the crew salaries make more sense and
be more immersive for the player. For example, a solo Interceptor pilot
(a highly dangerous role) now has a much higher salary than a random
crew member in the safety of a heavy warship. Conversely, huge ships
with tons of crew are now presumed to have a command structure. The
officers that run the ship receive a higher salary in recognition of
their rank and responsibilities.

I considered adding this logic to the `PlayerInfo` class, since that is
where crew salaries used to be calculated. However, `PlayerInfo` is a
really huge class with a ton of responsibilities. When editing it, I
often got lost if I moved my editor window too far. Instead, I added a
new class called `CrewEconomics`.

This commit does not include any UI changes, so the new salary amounts
may be confusingly high. Before we release the new calculation, we
should update the game content so that the new process makes sense to
players. Ideally the UI would display a salary amount for each ship.

Future Plans:

I would like to build on the concept of commanders and officers. I have
a system in mind where the player shares a portion of their profits with
the commanders and officers in their active fleet.

I also want to build a morale system, so that each ship has a morale
rating. Morale would be affected by various factors, including whether
or not the crew are paid, and could be 'repaired' by granting ships paid
shore leave (ie. parking them on a nice planet). A comment in the code
that I just added makes reference to this.