fix: Typo in overclocked repair module (#4705)
* Fixing Overclocked repair module spelling mistake

Line 160:
unparalled -> unparalleled

* Correcting sentence structure

Original: "providing unparalleled hull repair in all of Coalition space."

New: "providing hull repair unparalleled in all of Coalition space."

The former is stating that it provides unparalleled repair that works throughtout coalition space. The new version states that it provides hull repair that is the best in coalition space.