Minor style fixes and removed obsolete TODO
Style-fixes:
Renamed DoStep to Step (to match style of other code) and changed int to
enum TestResult (which is more specific and appropriate).

TODO-removal:
The inject mechanism for savegames writes the injected save-file directly
to the savegames directory. So no need to have two different load
mechanisms, just loading from the savegames directory is sufficient.

Other TODO was already reporting an error based on the RESULT_FAIL.