Adding the Subsidurial
This adds the Subsidurial, an indigenous space creature. It can be found in all uninhabited systems across the galaxy, but has a very, very low spaw rate. (60000).

It is intended to add a little interest to these otherwise lonesome places, in much the same way that the hyperion did in EVN.

Images courtesy of BeccaBunny