Update remnant missions.txt
Adds having completed 2 Broken Jump Drive missions as a potential alternative to completing the Keystone delivery.

This is intended to prevent situations such as the one that klarthailerion got himself into where he had someone managed to get Key Stones and thus access the Ember Waste, and had already met the Hai, and didn't have any keystones.

While the current variations on the Keystone mission shoudl resolve this problem; there is also the problem that the Keystone missions sound like off-hand side-quests for some extra credits; not essential-to-the-plot missions. Nor do they provide a warning after they have been declined/aborted that no, really, the player should go back and do them.

Instead of adding a 4th wall breaking "That was actually important, reload and do it" dialog, I think it is more fitting to allow helping with the Broken Jump Drive mission string to count in place of the keystone missions works better: It is providing rare technology to the Remnant; and the player has to put themselves at risk to do so (disabling Korath). This sounds like perfect qualifications for fulfilling the Prefect's statement about how "you have helped us in other ways as well."