Use Body class as formation-lead and use facing when standing still.
Changed the formation-positioner to use the Body-class as reference for the object around which to form a formation.
This should help if we want to make formations around non-ships in the future.

Also changed the formation-positioner to use the facing-property of the formation-lead when the formation-lead is not moving.
This makes the formations turn together with the lead when the lead is turning without moving.