Reducing the Outfit Size of Solar
This change reduces Solar panels to 1/3 of their former outfit space requirements - but leaves their mass. Thus a KP-6 panel takes up only a single ton of outfit space, but adds 3 to the mass of the ship. Which makes sense, considering that the description itself specifies that these are bolted to the outside of the ship. Not exactly the sort of thing that takes up much internal space.

The Coalition panels are treated similarly, except they are halved instead of reduced to a third.

Exact values likely subject to spreadsheeting.