The limits of Hai pacifism are tested after 900 Hai are captured by Human slavers. Captain <last> lacks a ground combat force, and must seek help from a government.  To benefit the greater good, the Hai Elders, Republic Parliament, and Syndicate refuse to intervene. That leaves only one group of unlikely heroes: the Unfettered Hai.
They were the protectors of the Hai once before. Can they be so again?
Is the divide among the Hai too great to cross? Or is there some hope
of peaceful reconciliation?

This story was inspired by an unpublished story by Arachi. It bears
little resemblance to the original, but the original story's spirit
remains strong.