The Unfettered Hai food situation grows desperate as the environment on Warfeed nears a catastrophic collapse. The Hai refuse to help terraform Warfeed in fear it would make the Unfettered too strong, and the Wanderer war effort is too slow to feed the masses. Tired of insulting food donations from the Hai, the Unfettered turn to Captain `<last>` for help setting up a large-scale smuggling operation.
This story drags you into the night side of endless sky, to marauders,
a Hai mafia, a Syndicate "off-shoot," crooked bounty hunters,
genetically modified humans, and even an insane six-year-old. As you
emerge, you're not sure you've done more good than harm.

The story is intended to be playable if you turn pirate. All stops
have clearance because everyone is asking you to do something illegal.

There is also a secret super-boss fight.

The main quest reward is a unique outfit; there is a short, follow-on,
Drak/Quarg, side-story to upgrade that outfit late in the game.