Removed "Navy" license from Carrier and Cruiser
These two ships have a higher license restriction - that of Carrier and Cruiser respectively. Given that these two licenses are both only given after a player has already received the Navy one, and are themselves Navy restricted licenses, there is no point in having two license restrictions instead of just listing the more restrictive of the two.

The only reason to keep this requirement is if there is going to be a situation wherein a player may acquire a carrier license without having a navy license. (which implies civilian carrier class ships that are restricted by this license), or if there is some way that a player who eventually earns a carrier license may lose their Navy one. (again, this implies the existance of other carrier class ships.)