Tweaking
Adjusted the total mass upwards a bit to reflect that it includes both the mass of two canons (minus a bit for efficiencies) plus the mass of the turret mount.

Revised the actual outfit space and weapon space down slightly, as this is aiming to be an unlocked weapon after a mission chain instead of a "available right away" sort of thing.

- Removed the turret from the outfitter. This will be unlocked in a mission chain.

- Removed the turret from the Pelican.