Update credits.txt
Shrunk the writers list a bit, and moved Jana over to the Editors list.

Removed "mybuddymichael" as it isn't actually a github account, but just a name attached to a commit by MZ.

Utilized MCO's full list to update the list of github contributors. (main change is that endless-sky and the devs are now on the list, which is fine.)