Tweaks to the Capture Ladder:
1. It takes about 800 tries for a Dredger to capture a Korath Freighter
2. Freighter can easily capture the other two old world-ships and the Quarg
3. Transport takes about 300 tries to capture a Refabricator

The bunk counts are lowered. I did this by adding a Korath Defense Station whose image I have not yet made.

It is nearly impossible for a Dredger to capture Quarg ships or the Korath Transport. (Takes more than 10000 tries.)

The Pug Arfecta is already within reach of an unmodified Bactrian (on average six tries).

This was determined using a monte carlo simulation based on the capture calculations in the code.