Added multiplier attributes to the whitelist
Values down to -1 are allowed. This means that at their lowest value, multipliers will have a -100% effect on the specified stat. They can not go lower then this, because then they would be flipping the sign on whatever the stat is, e.g. causing a shield generator to go from consuming energy to producing it.