Review changes
DrawFlareSprites now takes in the exact DrawList needed instead of the array, and all other parameters are const references.

Rearranged the if statement in DrawFlareSprites as to be more efficient.

EnginePoint now uses an Angle instead of a double.

Renamed facing to steering on EnginePoints to make it clear that this determines if the flare shows based off of the steering direction.

Renamed Angle() to Facing() on EnginePoints as to use the same vocabulary as Bays.

isReversing now checks to see if reverseThrust is even present.