Also consider "moving to position held" as HOLD position.
The HOLD command was not properly toggling for groups of ships that
already had received the HOLD_POSITION command, but where one or more
ships were (still) busy moving to the position that was to be held.

This commit adds a HOLD_ACTIVE command that handles the moving as
part of holding position to ensure that the HOLD command toggling can
differentiate between move-commands and move-commands that are part of
position-holding.