Adding Automata Costs
One of the main problems with the automata is that they break the game economy. Namely, they are free. Once you get a 512, you've got it for free for life.

This gives all automata ship a "maintnance cost" equivalent to their category:
- Heavy Warship: 1000
- Medium Warship: 600
- Light Warship: 250
- Interceptor: 50
- Fighter: 25
- Drone: remains free

Note that this is incredibly cheap compared to other faction's ships.  A fighter (typically 1 pilot, so 100/day is instead 25 a day.Heavy Warships range from 18 to 250 crew, which means 1800 to 25000 credits per day. Regardless, 1000/day is downright cheap, especially when it is a warship that outclasses most purchasable warships.