Remnant-Deep Additional Options
On the Deep side: This gives players who have already met the Remnant before doing the Deep mission the opportunity to *not* mention the Remnant to the Deep, and instead tell them about the wormhole train. It then leads to a mission where they scan the entirety of the wormhole train.

On the Remnant side, this introduces a small set of missions where the player has the opportunity to tell the Remnant about the research being done on the Ember Threshold wormhole, with an alternate mission if they have done the wormhole research for the Deep. (one path has them tracking down the researchers and finding what they published, the other enables the player to skip some steps and go straight to the home university for the results.