Skip jobs if the planet is not inhabited, not only if the planet does not have a spaceport
A planet is considered inhabited if it was a spaceport, or if it was a required reputation, or if it has a defense fleet, and if it lacks the uninhabited attribute. If a planet is inhabited, then it will show the jobs board when landed. But, jobs are not generated if no spaceport is present.

Instead check to see if the planet is inhabited to decide whether to skip generating jobs.