Re-created Hexagon formation using polar coordinates
The skip last option is not implemented yet, so there is overlap in the formation.
The centered option also is not implemented yet, the formation will look nicer once it is implemented.
Implementing this with polar coordinates and the skip option was a lot easier than manually calculating all positions without overlap.
Coordinates are based on ship-sizes and fixed positions, this should allow the scaling to work more nicely.