Outfitter helpl messages now display sequentially
Changed the order the code pulls the help messages: Now it will display each of the three help messages in sequence when the player visits the outfitter.

Unfortunately, it currently does so *every* time they visit the outfitter.

Note: It is still displaying all three messages simultaneously, but the first one is the longest and the third one the shortest, so you should never see the later ones while the earlier one is still on the screen. So effectively sequential but not very tidy.