Messages One at a time!
Now the messages only display when the outfitter is on top. So the messages don't stack.

So the counting has been reversed again to count up, since they aren't stacking and the player won't be seeing the last one first anymore.

So: Now the messages show from 1 to 3, one at a time.