Apply suggestions from code review
Thanks to Amazinite, Pointedstick, Anarchist2, and W1zrad for the suggestions.

General note: Putting explanations for changes with the code changes is much appreciated; but it would also be appreciated if general comments or suggestions that *I* change something be kept separate from the commitable code changes, because committing them causes GitHub to eat the rest of your comments. Or at least, hides them.

Co-authored-by: Amazinite <jsteck2000@gmail.com>
Co-authored-by: Anarchist2 <60202690+Anarchist2@users.noreply.github.com>
Co-authored-by: Wizrad <jbw3636@gmail.com>