Pushing what I have written so far
This reworks the mission structure and finishes the missions up until the player gets the technology to bring back to Hai space.

Still need to finish the conversation for when the player actually returns with the technology.

The Coalition half of this was written by Arachi. I haven't gone through it though, so will probably be tweaking some things with it in a later commit.

Co-Authored-By: Arachi <44789960+Arachi-Lover@users.noreply.github.com>