Downgraded to 5 and 4
People thought 8 and 6 was too high, so dropped down to 5 and 4.

The values are skewed that way because I perceive Alphas being a lot more lethal on the offense. Regular humans don't seem to have as much trouble attacking them as when they get attacked by them.

(Basically, the 5 represents a combinaton of physical abilities, experience, and mental abilities leading to better tactics that make them more effective)