Downgraded to 3-3
As per Amazinite's request for them to not be very strong. Assumption is that a large part of the crew is pirates and/or slaves.

I still feel that they need to be at least superior to the navy, though. If they had "more Alphas than we could imagine" at their base under the ice cap, it isn't beyond the realm of possibility that their flagship (and indeed, their *only* ship) might have had a mostly-Alpha crew.