Refactor part 2
DoWeather now uses shipCollisions to determine which ships to damage.

CollisionSet has been updated with a vector of all tracked bodies in order to do this. If this is undesired, we could instead create a vector of in-system ships that Engine keeps track of alongside shipCollisions. (That would avoid the reinterpret_cast<Ship *> that is currently done in DoWeather.)