Add BRANCH and LABEL and TODO items for test-timing.
The BRANCH and LABEL steps where added as replacement to make the test
code match the syntax of the conversation nodes more closely.

Also added TODO items that describe when the game takes a step forward.
This is required to ensure that the test can give continuous input
when desired (to ensure that the game only steps forward after inputs
or other events that requires the game to step).