Adjusting stopovers
Adjusting the text for each planet so that they do not mention the planet name.

Note: The engine currently does not appear to support multiple specific `on stopover` blocks. I'm leaving both in on hopes that someday it will, but in the meantime both blocks are written such that they can be triggered in either order on either world, and one on each world, without appearing to conflict.