Adding penguin branch
Adds a short branch to tell the player that they can use a penguin, but the rest of the fleet will still have to use puffins.

Also pluralized the on complete section so that it is clear that it is talking about the group of puffins; not the player's specific puffin.