Create navy missions intro.txt
Ladies and gentlemen...it is with great pomp and circumstance that I bring you something everyone's been waiting for...

The long awaited...

NAVY CAMPAIGN!!!

...Intro. ^_^;;;

We've still got a long way to go before the whole thing's ready, but you can freely take a look at what we've got so far.

More will come in the following months. Really, crossing this hurdle is the big one.