Added Cognizance 19
Swaps in the "inaccessible" tag for future compatibility with new and inaccessible systems.
Adds Cognizance 19 which is a short mission wherein Dawn asks the pilot to upload a copy of their Nav logs next time they visit a planet.
This triggers an event that removes the inaccessible tag from those systems.