fix(hazards): Estimate the distance traveled for the hazard
Blast radius weapons use this same estimation. Doing this allows hazards to dropoff to 0 damage and damage ships touching the hazard, whereas before if the ship's center was outside the hazard but the mask was touching the hazard then hit effects would be created without the ship actually taking damage.