Indigenous Ion Resistance
As organic entities that lack much of the electronic systems of conventional ships, the Archon, Subsidurial, Void Sprites, and Ember Waste Node shouldn't show nearly as much ionization effects as normal ships. They are basically immune to ionization anyway - this just reduces the visibility of ion effects by reducing how much of it is present while not actually eliminating it.