Rebalancing Inhibitor Slow Effect
Inhibitors currently do .5 slowing damage per shot. In testing, VitalChip found that with their modded values, it took 4 fully loaded Albatrosses (24 inhibitor cannons and 16 turrets) to almost stop a SB. (that was using values of .8 slow for canons and 1.6 for turrets)

Given tha the game is supposed to be balanced for solo fleets, this means that with the current value it is functionally impossible for a solo ship to inflict useful amounts of slowing on an enemy.

As such, this PR raises the "per shot" slowing damage value for the Inhibitors up to 2. This is still a very weak slowing effect, but should be enough that players can actually use it as a tactical weapon for controlling the flow of battle.