Revising the description
The second part of Amazinite's description is good; however the first part... Well, if a weapon has to *tell* you that it is formidable, maybe it isn't as good as it thinks it is?

I'd rather talk about the Remnant and how they came to have this weapon instead of something that sounds like it is trying to convince people that yes, it is actually good.