Remove unused teststep-type and unused communication enums
Teststep-type INVALID is no longer needed (test-status BROKEN handles all
situations where INVALID would have been used).

And the enum class Result that was used for communication between a
teststep and the test is no longer needed, since test-code now runs in
test (and the step is just a container).
This was based on the conversation class where the nodes were also
just containers and the Conversation class itself contains the actual code.