Trimming Planets (Making things dangerous)
As per Amazinite's request to cut down on the number of landable planets in Remnant systems, this adds the "requires: stabilizers"  attribute to six worlds.

In the same way that "gaslining" is a keyword for "Ships that have been designed to high-pressure hulls;" "stabilizers" is short for "Ships that have environmental stabilization systems to protect them when landing on hazardouse surfaces. Like floatation landing pads, etc"