Culling 11 visitable objects from Remnant space
Do to how the UI functions in-game, on small monitors the number of visitable planets in some Remnant systems (particularly in a gaslined ship) overloads the display and makes it impossible to see all the visitable places.

This PR adds the "requires: inaccessible" attribute to 11 such places and trims the Remnant systems down to 3 visitable planets:
- 1 inhabited planet (visitable normally)
- 1 uninhabited planet (visitable normally)
- 1 uninhabited gas giant (visitable with gaslining)

So normal ships will see 2 visitable planets in each system, and gaslined ships will see 3.