Showing the Black Diamond some lock
Black Diamonds are unique fighters that the player can never replace, so let's make them a bit more worthwhile.

Increased the HP of the fighter to be one of the stronger fighters in the game.

Gave the fighter some powerful hull repair built in at no cost aside from requiring that the fighter not be under fire in order to heal.

Increased the hull cost because it was rather low.

Buffed the damage of the Shard Fabricator by 100% and of the Heavy Ion Cyclotron by 50%.