fix(start-conditions): avoid creating ill-formed starting ships
We don't fully support "stock" ships as something that can be gifted to the player, but it works somewhat, in that attempting to give the player a named stock ship will give them the stock ship, but with no name. (The code parses it as though it were a variant definition.) If, however, no name was given, and no child nodes to define the ship were given, then the created ship would be completely empty.

 that is:
 ```
 start
   ship "Aerie" # very bad, player given an "Aerie" in name only
   ship "Aerie" "Evil Twin" # OK, player given a stock Aerie with no name, with an internal variant name of "Evil Twin"
```