Fix: Missing link in Remnant: Postverta Reveal event
The `Remnant: Postverta Reveal` event specifies links for several systems in the cluster. However, this specification replaces the system's old links instead of adding to them. In this context, the event is missing one link.

This adds the missing link to the event, and moves the postverta link to the end of the list so that they are in proper alphabetic order.

Fixes#5680