Apply suggestions from code review by W1zrad
Thank you for the Suggestions!

In regards to contractions, I think we decided that when the character is (or could be assumed to be) using sign, that they would not use contractions. While it is unclear, it can be assumed from what we see from the NPCs that either a) sign language doesn't have contractions, or b) the NPCs that the player deals with are spelling things out to avoid confusing someone who is just learning. If it is option A, then the player *can't* use contractions. If it is option B, then <first> won't learn contractions because no-one is demonstrating how they can be done. For our implementation purposes here, it amounts to the same thing.

My to-do list:
- Line 4613: replacing the double "reference"
- Line 4531: Consider doing a branch with a defer if the player doesn't have a Research Laboratory.

Co-authored-by: Wizrad <onlyw1zrad@gmail.com>