Replacing Main Page link to google groups with discord link
We really shouldn't be directing people to the google groups from the front page of the Endless Sky repository. None of the core developers or content creators frequent there, whereas they virtually all are on the discord. Let's direct people interested in contributing to where they can actually connect with us.